Scale/crop/frame the picture, for printing as photograph
The original picture, while it had the jpg extension, was apparently
in PNG format, since it was partly transparent, on the whole surface
of the picture. It started by converting the picture to JPG to remove
the partial transparency, then resized it to 15x10cm for printing as
a reference photograph for the project. I cropped the picture slightly
in so doing, removing a shadowed area at the top of the picture, and
adding white bands on the left and right to complete the frame.