Add extracts from Journals relating events around the death of Te Rakau
Also, improve formatting in page 24 (Te Maro).

I also decided to use the alternate picture as main picture,
because it is explicitly mentioned in the journal of Sydney
Parkinson as a reference picture for one of the warriors while
he relates the encounter leading to the death of Te Rakau.