Replace picture, reverting to the BnF as a source
I have worked on the scale and the balance of the colors.
The previous version of the picture was too cropped.
I started from the full page of the book, downloaded from Gallica,
to keep the frame around the map.

I have also added the previous/next navigation and the picture caption.