Clarify date of Cook's log, complete reference about date discrepancies
I now have an explanation for the difference between the date
in the journal of James Cook (using nautical dates) and the
journal of Joseph Banks (using civil dates).

The difference with the date used in Sydney Parkinson's journal,
the 9th early in the morning vs the 8th PM, remains mysterious.
The date would correspond to the Nautical Date, but the description
starts with the morning and omits the description of the evening before:

Civil Date: October 8th, PM
Nautical Date: October 9th, PM

is followed with:

Civil Date: October 9th, AM
Nautical Date: October 9th, AM